move localize and share types also fix some small stuff in the typings
