+++ src/main/types.d.ts
@@ -1,20 +1,17 @@
 import { App } from 'electron'
 
-export interface Login {
-  addr: string;
-}
+import { RC_Config } from './rc'
 
-export interface AppState {
-  deltachat: any;
-  logins: Array<Login>;
-  saved: boolean;
-}
+import { LocaleData } from '../shared/localize'
+
+import { AppState } from '../shared/shared-types'
 
 export interface ExtendedApp extends App{
-  rc: any;
+  rc: RC_Config;
   isQuitting: boolean;
   ipcReady: boolean;
-  localeData?: any
-  state?: any;
+  localeData?: LocaleData;
+  state?: AppState;
   saveState?: (arg: AppState) => void;
+  // saveState is likely not shared to renderer so the type ExtendedApp can not be shared with renderer
 }
